+++ src/parser/path/mod.rs
@@ -3,11 +3,10 @@ mod regex;
 
 use std::vec::Vec;
 
-use crate::document::Document;
-use crate::error::*;
-
 use super::PathCandidate;
 use super::tree_sitter;
+use crate::document::Document;
+use crate::error::*;
 
 pub fn parse_document(document: &Document) -> PathServerResult<Vec<Vec<PathCandidate>>> {
     Ok(extract_string(document)?
@@ -19,12 +18,13 @@ pub fn parse_document(document: &Document) -> PathServerResult<Vec<Vec<PathCandi
 /// Extract string tokens from the document
 fn extract_string(document: &Document) -> PathServerResult<Vec<PathCandidate>> {
     let res = tree_sitter::extract_strings(document)?;
-    if let Some(res) = res {
-        Ok(res)
+    let res = if let Some(res) = res {
+        res
     } else {
         // fall back to general parser
-        Ok(regex::extract_string(document).unwrap_or_default())
-    }
+        regex::extract_string(document).unwrap_or_default()
+    };
+    Ok(res)
 }
 
 /// Try to extract paths from a string token,
@@ -39,12 +39,7 @@ fn extract_paths_from_string(path_ref: PathCandidate) -> Vec<PathCandidate> {
     }
 
     // Level 2: the part of string (split by space) is a path or not
-    results.extend(
-        path_ref
-            .split(&[' ', '\n'])
-            .into_iter()
-            .filter(|part| !(part.content == "/" || part.content == "\\")), // drop single slash
-    );
+    results.extend(path_ref.split(&[' ', '\n']));
 
     results
 }
@@ -118,15 +113,5 @@ mod tests {
             eprintln!("Extracted: {};", p.content);
         }
         assert!(res.iter().any(|p| p.content.trim() == "/tmp/目录/"));
-    }
-    #[test]
-    fn test_filter_slash() {
-        let candidate = PathCandidate {
-            content: "a / b".to_string(),
-            start_byte: 0,
-            end_byte: 1,
-        };
-        let res = extract_paths_from_string(candidate);
-        assert!(!res.into_iter().map(|p| p.content).any(|c| c == "/"));
     }
 }